Automate download of library files
This commit adds automatic downloading of all required files for the
current callers. It also moves the default download location to be a
directory above the output directory; in this way, multiple runs can use
the same set of library files, even if the checkpoint files have been
lost. This commit also introduces some minor name changing for
consistency.